feat: colour picker
Currently only set it for pen, but tried to make it generic enough to be
used anywhere
When the colour bubble is clicked, a little menu appears above with a
few preset colours. Once one is clicked, that is the active colour.
Either the bubble can be clicked again or drawing triggers the popup to
close.