feat: added typescript
Added typescript primarily for the nice tooling.
At this point of my life I'm using Vim (neovim) with lsps, but for JS
the LSP is tsserver, which is for TS, so without it all of my goto
functionality is moot.
Yes this is a small codebase but still it's nice to have that
functionality.